:bug: Fix chatbot use the correct context loader. - Moved `GetClientByPhone` use case class from `src/api/cases/client_use_cases.py` -> `src/api/saragurosnet/cases/client_cases.py`. - Removed `GET` `/api/v1/clients/{client_phone}` endpoint. - Actions in tree use correct use case.
