+++ src/api/cases/__init__.py
@@ -1,14 +1,9 @@
 from src.api.cases.auth_use_cases import LoginUser, RegisterNewUser
-from src.api.cases.client_use_cases import (
-    GetClientByPhone,
-    GetClients,
-    RegisterNewClient,
-)
+from src.api.cases.client_use_cases import GetClients, RegisterNewClient
 from src.api.cases.message_use_cases import get_last_messages, write_message
 
 __all__ = [
     "get_last_messages",
-    "GetClientByPhone",
     "GetClients",
     "LoginUser",
     "RegisterNewClient",
