+++ src/api/cases/client_use_cases.py
@@ -23,10 +23,10 @@ class GetClients(UseCaseBase):
         return create_response(clients, "Clients retrieved")
 
 
-class GetClientByPhone(UseCaseBase):
-    def __call__(self, client_phone: str):
-        with self._session() as session:
-            client_repository = ClientRepository(session)
-            client = client_repository.get_client_by_phone(client_phone)
+# class GetClientByPhone(UseCaseBase):
+#     def __call__(self, client_phone: str):
+#         with self._session() as session:
+#             client_repository = ClientRepository(session)
+#             client = client_repository.get_client_by_phone(client_phone)
 
-        return create_response(client, "Client retrieved")
+#         return create_response(client, "Client retrieved")
