+++ src/saragurosnet/cases/__init__.py
@@ -1,0 +1,5 @@
+from src.saragurosnet.cases.client_cases import GetClientByPhone
+
+__all__ = [
+    "GetClientByPhone",
+]
